Delete launch.json, add task.json
Added task.json and made it work with bash shell. As for now, I
don't know how to make launch.json work with bash shell, so I
temporarily deleted it. One of the problems is that it makes new
terminal every time it executes. For now, I have to manually type
to execute it in bash.